Working on Alexandria Manual.
Split alexandria.xml into multiple files to make it easier to work
with, there is now one file per chapter.

Applied GNU Free Documentation License v1.2 (or later) to the
manual. Thanks to Liam Davison for agreeing to this (previous to this,
the manual was not strictly free documentation, although that was his
intent).

Removed the old screenshots, since the GUI has changed considerably
since 0.6.1 when they were taken. Added a new set of screenshots taken
from the new 0.6.2 GUI.

Added details about Smart Libraries (a new feature since 0.6.1).

Made a few changes to refer to new or changed behaviour.


git-svn-id: http://alexandria.rubyforge.org/svn/trunk/alexandria@814 9992d806-23ff-4cc5-b4c6-abbc17c6a14e