Switched old LS provider for new Siciliano provider. Also fixes #9008, so it can now be advanced in the preferences order.
Requires Hpricot for HTML parsing (this is a very tricky provider, I
wish it were possible to find an easier-to-use alternative). It
requires a work-around to search for both ISBN-13 and ISBN-10 in turn,
since the website doesn't translate between them easily. There is some
JavaScript trickery with the cover image, so we get a smaller one than
displayed on the website. There is no method for determining the URL
from an ISBN (the permalink uses a product code).


git-svn-id: http://alexandria.rubyforge.org/svn/trunk/alexandria@1004 9992d806-23ff-4cc5-b4c6-abbc17c6a14e